:art: Ausbilder Data into component
